pyproject.toml
[project]
name = "siq-vl"
version = "0.1.0"
description = "Seek Vision Language Model"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.4.1",
    "deepspeed>=0.18.2",
    "pillow>=12.0.0",
    "setuptools>=80.9.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wandb>=0.23.0",
]

[tool.setuptools]
packages = ["siq_vl"]


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
